requirements.txt
accelerate
boto3
datasets
ensure
evaluate
faiss-cpu
fastapi
huggingface_hub
Jinja2
langchain
langchain_community
langchain_groq
langchain_huggingface
matplotlib
mypy-boto3-s3
nltk
notebook
pandas
protobuf
py7zr
pypdf
python-box
python-dotenv
PyYAML
rouge-score
sacrebleu
sentence-transformers
sentencepiece
streamlit
tqdm
transformers
uvicorn
torch
rank_bm25
qdrant-client 
langchain-qdrant
sqlalchemy 
asyncpg 
alembic 
redis 
passlib[bcrypt] 
python-jose[cryptography] 
python-multipart